Image handler
This is the view of the new subsection, localmap, which displays an image/warning provided by the controller and crudely adds a form and iframe to handle the asynchronous file upload. Due to the conflicting form styles, this cannot use the BuildOutput form automation, and instead the code appears here in the raw.